Move specialized optimization logic from SegmentedArray to SegmentedList usage point.
