+++ src/Compilers/Core/CodeAnalysisTest/Collections/List/SegmentedList.Generic.Tests.cs
@@ -73,88 +73,5 @@ namespace Microsoft.CodeAnalysis.UnitTests.Collections
             Assert.Throws<ArgumentException>(null, () => list.CopyTo(0, new T[0], 0, count + 1));
             Assert.Throws<ArgumentException>(null, () => list.CopyTo(count, new T[0], 0, 1));
         }
-
-        [Theory]
-        [MemberData(nameof(ValidCollectionSizes))]
-        public void Capacity_ArgumentValidity(int count)
-        {
-            var list = new SegmentedList<T>(count);
-            for (var i = 0; i < count; i++)
-                list.Add(CreateT(i));
-
-            Assert.Throws<ArgumentOutOfRangeException>(() => list.Capacity = count - 1);
-        }
-
-        [Theory]
-        [InlineData(0, 0, 1)]
-        [InlineData(0, 0, 10)]
-        [InlineData(4, 4, 6)]
-        [InlineData(4, 4, 10)]
-        [InlineData(4, 4, 100_000)]
-        public void Capacity_MatchesSizeRequested(int initialCapacity, int initialSize, int requestedCapacity)
-        {
-            var list = new SegmentedList<T>(initialCapacity);
-
-            for (var i = 0; i < initialSize; i++)
-                list.Add(CreateT(i));
-
-            list.Capacity = requestedCapacity;
-
-            Assert.Equal(requestedCapacity, list.Capacity);
-        }
-
-        [Theory]
-        [InlineData(0, 0, 1, 4)]
-        [InlineData(0, 0, 10, 10)]
-        [InlineData(4, 4, 6, 8)]
-        [InlineData(4, 4, 10, 10)]
-        public void EnsureCapacity_ResizesAppropriately(int initialCapacity, int initialSize, int requestedCapacity, int expectedCapacity)
-        {
-            var list = new SegmentedList<T>(initialCapacity);
-
-            for (var i = 0; i < initialSize; i++)
-                list.Add(CreateT(i));
-
-            list.EnsureCapacity(requestedCapacity);
-
-            Assert.Equal(expectedCapacity, list.Capacity);
-        }
-
-        [Theory]
-        [InlineData(1)]
-        [InlineData(2)]
-        [InlineData(4)]
-        public void EnsureCapacity_GrowsBySegment(int segmentCount)
-        {
-            var elementCount = SegmentedArray<T>.TestAccessor.SegmentSize * segmentCount;
-            var list = new SegmentedList<T>(elementCount);
-
-            for (var i = 0; i < elementCount; i++)
-                list.Add(CreateT(i));
-
-            Assert.Equal(elementCount, list.Capacity);
-
-            list.EnsureCapacity(elementCount + 1);
-            Assert.Equal(elementCount + SegmentedArray<T>.TestAccessor.SegmentSize, list.Capacity);
-        }
-
-        [Theory]
-        [InlineData(1)]
-        [InlineData(2)]
-        [InlineData(4)]
-        public void EnsureCapacity_MatchesSizeWithLargeCapacityRequest(int segmentCount)
-        {
-            var elementCount = SegmentedArray<T>.TestAccessor.SegmentSize * segmentCount;
-            var list = new SegmentedList<T>(elementCount);
-
-            for (var i = 0; i < elementCount; i++)
-                list.Add(CreateT(i));
-
-            Assert.Equal(elementCount, list.Capacity);
-
-            var requestedCapacity = 2 * elementCount + 10;
-            list.EnsureCapacity(requestedCapacity);
-            Assert.Equal(requestedCapacity, list.Capacity);
-        }
     }
 }
