+++ src/Dependencies/Collections/SegmentedArray`1+PrivateMarshal.cs
@@ -2,6 +2,8 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
+using System;
+
 namespace Microsoft.CodeAnalysis.Collections;
 
 internal readonly partial struct SegmentedArray<T>
@@ -14,5 +16,17 @@ internal readonly partial struct SegmentedArray<T>
         /// <inheritdoc cref="SegmentedCollectionsMarshal.AsSegments{T}(SegmentedArray{T})"/>
         public static T[][] AsSegments(SegmentedArray<T> array)
             => array._items;
+
+        public static SegmentedArray<T> AsSegmentedArray(T[][] segments)
+        {
+            if (segments is null)
+                throw new ArgumentNullException(nameof(segments));
+
+            var length = 0;
+            foreach (var segment in segments)
+                length += segment.Length;
+
+            return new SegmentedArray<T>(length, segments);
+        }
     }
 }
