+++ src/Dependencies/Collections/SegmentedCollectionsMarshal.cs
@@ -23,6 +23,24 @@ internal static class SegmentedCollectionsMarshal
         => SegmentedArray<T>.PrivateMarshal.AsSegments(array);
 
     /// <summary>
+    /// Gets a <see cref="SegmentedArray{T}"/> value wrapping the input T[][].
+    /// </summary>
+    /// <typeparam name="T">The type of elements in the input.</typeparam>
+    /// <param name="segments">The input array to wrap in the returned <see cref="SegmentedArray{T}"/> value.</param>
+    /// <returns>A <see cref="SegmentedArray{T}"/> value wrapping <paramref name="segments"/>.</returns>
+    /// <remarks>
+    /// <para>
+    /// When using this method, callers should take extra care to ensure that they're the sole owners of the input
+    /// array, and that it won't be modified once the returned <see cref="SegmentedArray{T}"/> value starts
+    /// being used. Doing so might cause undefined behavior in code paths which don't expect the contents of a given
+    /// <see cref="SegmentedArray{T}"/> values to change outside their control.
+    /// </para>
+    /// </remarks>
+    /// <exception cref="System.ArgumentNullException">Thrown when <paramref name="segments"/> is <see langword="null"/></exception>
+    public static SegmentedArray<T> AsSegmentedArray<T>(T[][] segments)
+        => SegmentedArray<T>.PrivateMarshal.AsSegmentedArray(segments);
+
+    /// <summary>
     /// Gets either a ref to a <typeparamref name="TValue"/> in the <see cref="SegmentedDictionary{TKey, TValue}"/> or a
     /// ref null if it does not exist in the <paramref name="dictionary"/>.
     /// </summary>
